Restructure the code so it works better for day 1 and day 2 operations
