+++ src/components/common/GroupsApplicationsSelector/logForwardingGroupTreeData.ts
